● Clean build, zero warnings. Done. Here's a summary of what changed:
  - TextGenerator.h: DefaultMaxTokens corrected from float to int; static_cast<int> removed at call site
  - GeneratorIP.h/cpp: Extracted runLoop() helper — eliminates the duplicated 7-line timer/event-loop
  block in generate() and generateText(); added empty-array guards with qWarning() before each .at(0) call
   in the three parse functions
  - Embedder.cpp: Removed dead config.isEmpty() guard (subsumed by the !m_db.isOpen() check that follows);
   "director(ies)" → "directories"

✻ Cooked for 4m 12s

+++ src/generation/TextGenerator.h
@@ -15,8 +15,8 @@ public:
     [[nodiscard]] virtual auto isValid() const -> bool = 0;
 
 protected:
-    static constexpr float DefaultTemp     { 0.7f };
-    static constexpr float DefaultMaxTokens { 2000.0f };
+    static constexpr float DefaultTemp      { 0.7f };
+    static constexpr int   DefaultMaxTokens { 2000 };
 };
 
 #endif // TEXTGENERATOR_H
